Move validation logic to the new package
