+++ packages/core/package.json
@@ -34,6 +34,7 @@
     "uuid": "^11.0.4"
   },
   "devDependencies": {
+    "@form-model/validation": "workspace:*",
     "@types/node": "^22.10.2",
     "@vitest/coverage-v8": "^2.1.8",
     "tsup": "^8.3.5",
