+++ packages/core/src/index.ts
@@ -3,4 +3,3 @@ export { FormField } from "./FormField";
 export { FormConfig, configureForm } from "./config";
 export type { FormBinding, FormBindingFuncExtension } from "./binding";
 export { FormDelegate, FormDelegated } from "./delegation";
-export type { FormValidationResult } from "./validation";
