+++ packages/validation/src/index.ts
@@ -1,1 +1,4 @@
 export { getWatcher, watch, unwatch } from "./watcher";
+export { Watcher } from "./watcher";
+export { Validator } from "./validator";
+export type { FormValidatorResult, ErrorMap } from "./error";
